Move functions which require dependency on localevent.cpp/.h to localevent.cpp/.h
